utils: Add label module to overdraw to areas of the terminal.
A set of text labels can be output to the terminal then their contents
can be updated and re-output to the terminal in place. If output is not
directed to a terminal, the in place overdraw aspect is disabled and the
labels are just output normally.

This is not general purpose, but is good enough for our needs.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>