core: Abstract runtime to allow executing commands in a container.
Reuse the infrastructure used by tuxmake to abstract execution of
external subprocesses into a Runtime class. The required runtime can be
specified on the CLI. 'null' just executes on the native host and is the
default. 'docker', 'docker-local', 'podman' and 'podman-local' can be
used to execute within a container.

The container image can be specified on the CLI, and defaults to
'shrinkwrap/base-debian11', the official shrinkwrap image on dockerhub.

Our runtime wrapper class has a context manager to tidy up after itself,
and allows the user to request a final output command for a given input
command. This output command is a list that can be passed to
subprocess.* functions, as would have been previously done.

The runtime needs access to certain files/directories on the host. This
is done via appropriate bind-mounts.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>