config: Switch to yaml instead of json.
yaml is easier to format for human readability (e.g. strings can be
broken across lines, its less strict on quotes and commas, integers can
be binary, octal, decimal or hex, and comments can be inserted).

So switch over to using yaml instead of json. This adds a dependency on
the pyyaml package, since yaml is not part of the python standard
library.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>