run: Quit when fvp process terminates.
Previously, we would wait for all processes (fvp and nc instances) to
terminate before quitting the wrapper. But the interactive instance of
nc does not terminate immediately when the fvp terminates. Instead the
user has to feed a character of input to nc before it notices that the
fvp has gone and terminates.

So add a new parameter to the Process wrapper called `run_to_end`. If
true, the wrapper will not exit until the process has terminated. We
create the fvp process with run_to_end=True and all others with
run_to_end=False. So once the fvp terminates, the wrapper quits without
waiting for the nc instances (and will focibly kill them).

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>