config: Allow multiple git repos per component.
EDK2 requires 2 git repos to be independently cloned. So allow
expressing multiple repos.

The default descriptor does not change: a dictionary containing `remote`
and `revision` keys, who's values are both strings. However, this form
will now be normalized to the new canonical form: a dictionary of
disctionaries, where the key for the first level is the local location
where the repo should be placed (relative to the component path) and the
second level is the dictionary from the first form.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>