build: Save per-config build.sh scripts.
Previously we saved a global build.sh in the root of the package, which
contained the build commands for all built configs. However, this
prevents building configs consecutively since the script is overwritten
on each invocation.

Solve this by saving a per-config build script that only details the
steps for that config and save it in
<SHRINKWRAP_PACKAGE>/<config_name>/build.sh.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>