build: Improve parallelism and verbose logging.
Improve the parallelism so that new fragments that become ready as a
result of a previous fragment completing get started immediately if
there are free slots. Previously all ready fragments would need to
complete before we looked for new ready fragments.

As a side effect of this change, we can take advantage of ProcessManager
and its logging capabilities, so we can output marked up logs in
realtime.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>