build: Add --jobs argument to control max jobs in low-level build task.
We now have --tasks to control high-level parallelism and --jobs to
control low-level parallelism.

To hook this up to the config declarations, we added new infrastructure
to supply a 'clivars' dictionary. Each entry in this dictionary gets
exposed to the configs as a ${param:*} macro. So ${param:jobs} provides
the max number of parallel jobs that should be used. Where the
component's build system supports parallelism, this parameter should be
passed to it.

Some components currently have dependency bugs in their makefiles (tfa
and tfa-tests) so are hardcoded to always 1 job. Hafnium always uses all
cpus and there is no way to override that.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>