build: Ensure that cleanup always happens and only error is reported.
If there are parallel tasks running when a task fails, previously not
all of the running tasks would be correctly killed and cleaned up. This
could lead to script fragments being left in temp directories within
<SHRINKWRAP_BUILD> directory.

This is caused because when a fragment fails, any others are focibly
killed. The forcible killing causes them to report a non-zero exit code,
which causes a secondary exception to be raised. Fix by ignoring
exceptions in the cleanup path.

This has a useful side-effect of also ensuring we only output the logs
from the original failing task, and not any of the tasks that were
subsequently killed.

Original issue reported by Andre Przywara.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>