build: Don't attempt to build incomplete git sync.
Previously, we checked for the existance of the directory into which git
should have cloned the repo. If the directory existed then we assumed we
had done a git sync on a previous invocation (and optionally the user
may have made changes) so we did not attempt to sync the repo again.
This breaks if on that previous invocation, the git sync task did not
complete for some reason (ctrl+c, another parallel task failed, etc). In
that case the directory exists but the contents are not complete.

Fix this by adding a sync file. Its precense on the filesystem tells
shrinkwrap that a git sync is (was) in progress. If found, shrinkwrap
will destroy any existing repo and start the git sync again.

Original issue reported by Andre Przywara.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>