build: Sensible defaults for --jobs and --tasks.
Previously set default was 4 for each. Since there are 2 levels, it is
difficult to set a sensible default, but 4 is almost certainly not it.
Default to half the number of available CPU cores, clamped to 32. In
practice there are rarely more than a couple of active tasks at a time.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>