config: Make ${params:} macro more flexibe.
Previously, ${params:} was a macro that had a type (`params`) but no
name. And it expanded the appropriate params dictionary to a string of
`key=value` pairs.

Change this behaviour so that the type is singular (`param`) to better
align with other types (we are going to add scalars into this namespace
in future too). And use the name portion of the macro to describe the
separator that should be used. ${param:join_equal} will expand to
`key=value` whereas ${param:join_space} will resolve to `key value`. The
old ${params:} style is no longer valid and all configs are updated.

EDK2 requires the ${param:join_space} variant.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>