config: Rename config files to have .yaml extensions.
Add the .yaml extension to all the config files to make it clear to the
user that this is the format of the file. It also has the added benefits
that IDEs can automatically format them nicely and it means that we no
longer need the ugly "_artifacts" suffix on the config package
directories since there is no clash with the config file name.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>