config: Separate out source and build trees.
This is the first step towards being able to provide your own source
tree for shrinkwrap to build instead of forcing shrinkwrap to sync the
source from git.

Introduce two new macros, ${param:sourcedir} and ${param:builddir},
which configs can use to construct paths within source and build trees
respectively. We scrap the config `buildroot` key and expect that
artifacts are addressed using the new macros.

Where components support building with a build directory separate from
the source, the build commands are modified to specify this. (this is
all cases currently). If a component cannot support this then it will
still build into its source directory. But this is discouraged since it
means when supplying your own tree, it will get modified.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>