config: Allow specifying custom locations for source and build trees.
As of the previous commit, a component has two new keys; sourcedir and
builddir. These were previously just an output from shrinkwrap to
document where those locations are. With this patch, these are now
inputs too. If either are present, they will be used as the location for
source and build respectively.

The main intended use case is when a user already has a repo checked out
on their system for one of the components and they want to use that to
build from. But others are also now possible (e.g. pass a non-existant
or empty directory for sourcedir and shrinkwrap will do a git sync to
it).

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>