utils: Expose primary IP address of runtime.
Add ip_address() function to Runtime class, which wil return the primary
IP address of runtime. If the runtime is in a container, it returns the
IP address assigned to the container.

This will be needed for the upcoming change to allow the user to specify
how they connect to the FVP uart terminals. For `telnet` terminal, the
user needs to be told the IP address and port to connect to.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>