run: Introduce mechanism for executing commands prior to starting FVP.
We are about to make some changes to the EDK2 config that requires
generating a startup.nsh script on the fly just before starting the FVP.
To enable that, introduce the 'prerun' key to the 'run' section of a
config, which contains a list of commands to execute.

We also introduce a new macro, ${param:packagedir}, which is useful for
prerun commands to know where to generate any output.

To make this all work, we must now generate a temp bash script
containing all these commands to execute, in the same way that the build
command does.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>