config: Split edk2 config in two; acpi and dt variants.
EDK2 supports semihosting, so we use this to allow the user to
(optionally) pass kernel and command line (and dtb for the -dt variant)
at run-time. We then create a startup.nsh and boot the FVP so that UEFI
can see the files on the host.

This is split into two variants; ns-edk2-acpi.yaml boots the kernel with
ACPI, which ns-edk2-dt.yaml boots the kernel with a supplied device
tree, which defaults to the upstream kernel dt if not explicitly
provided).

See the descriptions in ns-edk2-acpi.yaml and ns-edk2-dt.yaml for
details.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>