config: Add ns-preload.yaml, the simplest possible config.
This preloads TF-A's BL31, a kernel and a device tree into FVP memory
and boots it. Control is provided for swapping out each component and
for customizing the kernel command line. If you just want to boot a
kernel on FVP as quickly as possible, this is what you need!

See description in ns-preload.yaml for details.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>