build, inspect: Filter configs with by 'concrete' flag.
Introduce a top-level flag into a config, `concrete` which defaults to
false. If true, it indicates that the config is intended to be built and
run directly (as apposed to being a fragment for inclusion in other
configs). This is intended to make things clearer for the user.

Enhance the inspect command to display this flag in the summary and also
to filter on it if no explicit configs were requested. The filtering can
be disabled with the --all option. No filtering is done if config names
are passed explicitly.

Enhance the build command to filter on it if no explicit configs were
requested, so that only the concrete standard configs are built. No
filtering is done if config names are passed explicitly.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>