build: Only constrain concrete set if implicitly building config store.
If the user specifies a config explicitly on the command line, we should
always attempt to build it, even if it is not marked as concrete. Only
filter to the set of concrete configs when the user doesn't pass any
configs on the command line, expecting to the build all the configs in
the config store.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>