build: Remove requirement that all bash variables are bound.
In hindsight this was a mistake. We were already having to work around
it for the EDK2 build, since it has scripts that reference unbound
variables that need to be sourced. It is also a problem where we are
using $CROSS_COMPILE, which is not defined when executing natively on an
arm64 system.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>