run: Bind-mount scripts into runtime.
Fix issue where bash script with the FVP commands was not correctly
bind-mounted into the container. It was working by accident because by
default the path where the script was created was under the current
directory, which is always bind-mounted. But changing SHRINKWRAP_PACKAGE
and SHRINKWRAP_BUILD expose the issue.

Fix this by explicitly bind-mounting the package directory, and create
the script in there rather than the build directory. `run` has nothing
to do with building so we should not touch that directory.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>