config: Add linux kernel build config.
This builds the kernel image ('Image') along with all modules (which are
archived into a .tgz) and both items are exported as artifacts.

By default the defconfig is built, but the user can overlay some
prebuild commands to modify the defconfig or replace it with a custom
one.

To mix this into a pre-existing firmware config (ns-preload.yaml shown
here), do the following (if you want to modify the config, you can
define your own yaml and pull in linux-base.yaml as a layer):

shrinkwrap -R docker build \
    -o linux-base.yaml \
    ns-preload.yaml

shrinkwrap -R docker run \
    -r KERNEL=\${artifact:KERNEL} \
    -o linux-base.yaml \
    ns-preload.yaml

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>