config: Auto-boot kernel for ns-edk2-* configs.
Use a pre-canned flash image which contains the UEFI variables to
auto-boot to the shell. This causes the shell to auto-execute
startup.nsh, which boots the kernel through semihosting.

Now that we can do this automatically, we bypass the UEFI UI app, which
was the main perpetrator of ncurses-like output. So we can just mux the
output to stdout rather than forcing the user to start a telnet session.
The shell still outputs a small amount of ansi escape codes, but we
clean that up before outputting.

edk2-flash.img was generated by configuring the model with:

-C bp.flashloader1.fnameWrite: edk2-flash.img

Then executing the following at the UEFI Shell:

$ fs0:
$ bcfg boot add 0 shell.efi "shell"

Then shutting down the FVP cleanly with ctrl+C.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>