config: Always boot FVP with SVE plugin.
This is the first plugin we have enabled, which meant we had to figure
out how to get the location of the standard FVP plugins. It turns out
that I had preemptively added the plugins directory to the path within
the docker images. So its just a case of using $(which <plugin name>) to
resolve the path at runtime.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>