run: Overhaul user input handling.
User input handling is overhauled so that it is now possible to use bash
within the FVP naturally. No more line buffering and 2x echoing. All key
combinations are passed through unmodified (including tabs, '^C', etc)
to allow tab completion and interrupting processes within running within
the FVP.

This is achieved by setting the input side of the user's terminal into
(mostly) raw mode and passing all input to a pty which is connected to
telnet, which is now used as the uart client for stdinout uart
terminals.

As a side-effect of passing '^C' through to the FVP, it can no longer be
used to quit shrinkwrap. Instead the '^]' key binding is used to quit
shrinkwrap. This is the only key combination not passed through.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>