config: Implement optional clean commands.
Introduce a new key, `clean`, in the component dictionary, who's value
is a list of commands to clean the component. This will be used by a
future patch that introduces the ability to clean components.

Not all components need to provide clean commands; sometimes it is
sufficient for their build directory to be removed.

As well as introducing the new key, we add clean commands to the configs
that require them.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>