clean: Add new command to clean components.
Some components (TF-A being a prime example) don't behave well when
changing their build parameters, without doing an intermediate clean.
This can cause issues when an overlay is changed, for example.

Introduce a new command, `clean` which behaves similarly to `build`
except that it cleans the target configs/components. Cleaning a
component consists of running some optional commands from the component
definition in the config, removing the component's build directory, and
optionally cleaning and resetting the git repo(s).

The clean command also adds a --filter option to allow cleaning only a
subset of the components.

The implementation reuses the same infrastructure as used by build.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>