build, clean: Keep run-time overlay config passed at build-time.
Given build- and run- time configs are interdependent, I think it is
smarter to keep hold of any run-time config that is passed at build-time
and pass it into the config that is output to the package (which is the
input config for the run command).

An example of where this behaviour is useful is using an overlay to tune
the target architecture version. There are tightly coupled changes to
both tfa (build-time) and the fvp (run-time).

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>