inspect: Parse config store tree.
The config store is now a tree with subdirectories that can store more
configs. So teach inspect to walk the tree and output all configs.

It is important that inspect outputs a config's name relative to the
root of the config store, so that users can easily see how to include or
invoke those files. So do some work to ensure the relative path is
maintained.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>