config: Require each component to specify their toolchain.
Add a 'toolchain' key to the component dictionary, which specifies the
toolchain to use. This string is set in the CROSS_COMPILE environment
variable before invoking any prebuild, build, postbuild or clean
commands.

When using the standard docker runtime image, the following options are
supported. Users are free to configure their own thing when using the
null runtime.

* aarch64-none-elf-    (arm64 bare-metal, from Arm)
* arm-none-eabi-       (arm32 bare-metal, from Arm)
* aarch64-linux-gnu-   (arm64 linux, from Debian)
* arm-linux-gnueabihf- (arm32 linux, from Debian)

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>