config: Support running multiple instances of ns-edk2-* in parallel.
ns-edk2-* previously relied upon assembling all the assets required for
semihosting in the package directory. Some of these assets were
dynamically generated at runtime. This was problematic for running
multiple instances in parallel.

So instead, create a per-instance temporary directory for semihosting
that is cleaned up at exit.

This approach means that we no longer need to expose the
{param:packagedir} macro, so remove that while we are at it.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>