docker: Add scripts to build and publish docker images.
To build the docker images do the following:

  cd docker
  ./build.sh latest

This will build and push the images for your native architecture (works
for both x86_64 and aarch64). Images are currently pushed to Arm's
internal Artifactory repo, and you will need to have docker logged in
for this to be successful. (We plan to change this in future to publish
to docker.io).

You will need to build on both x86_64 and aarch64, then do this on one
of the systems:

  ./publish.sh latest

This will tie the images together to form multiarch images, such that
docker can resolve the correct image based on the arch on which you are
running.

The following images are built:

  - base-slim-nofvp
  - base-slim
  - base-full-nofvp
  - base-full

'slim' variants don't contain aarch32 toolchains. These are massive and
bloat the image and most people don't need them. 'full' variants are
based on 'slim' but add in the aarch32 toolchains. 'nofvp' variants
don't contain an fvp. These images are intended for use as base images
if the need arises to add a custom fvp.

Currently there is no public fvp support for aarch64, so there is no fvp
included with this arch. We hope to solve this in future.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>