config: Improve dt-base.yaml to support user-defined DTS.
A higher layer config can now optionally override the DTS that gets
compiled. While we are at it, we also change the method of injecting the
kernel command line to be more robust (the previous method broke if the
chosen{} node was not already present or was in an included file).

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>