build, clean: Output warnings by default.
Modify the behavior so that even when --verbose is not specified, stderr
is still output to the console. This means that compiler warnings and
the like will be seen more clearly. In the absense of warnings, no
stderr is expected to be output so the console still looks "pretty",
with overdrawn status labels etc. If a warning occurs it is output
consicely and labels are reset so they don't overdraw the warning.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>