config: Use sensible default paths for config store.
Requiring the user to explicitly set the SHRINKWRAP_CONFIG environment
variable is onerous and error-prone. So pre-populate the path list with
common locations and prepend with any (now optionally) supplied paths
from the user. And filter so that we only keep paths that exist, which
also resolves a pre-existing bug where shrinkwrap errored out if
SHRINKWRAP_CONFIG contained a non-existing path.

While we are at it, print the final workspace paths when --verbose is
specified.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>