docker: Updates for storing images in Docker Hub.
Switch from using Arm's internal Artifactory repo to using the public
docker hub to store images. We also take the opportunity to modify the
naming scheme so that the arch is part of the tag instead of part of the
repository so that all images live in the same repository.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>