core: Add support for Python 3.6.9
Previously we were restricting support to Python 3.7.0 and newer, due to
that version being the first to support ordered dicts, which are relied
upon to pretify generated config files. However it turns out that Python
3.6 implements ordered dicts despite that version of the language not
requiring it.

Ubuntu 18.04 ships with Python 3.6.9 by default and a number of people
are still using that release, so let's support that as the minimum
version.

In order for it to work, there are a few mods we have to make to API
calls etc.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>