runtime: Add support for podman containers
As an alternative to Docker, users can now choose to use
--runtime=podman. It uses the same container images (from Docker hub)
but some people prefer podman because it doesn't require the use of a
damemon with root privilege.

Tuxmake provides most of the infrastructure so we just wire it up to the
--runtime cli option and fix a few issues where podman is more picky
(mount points and fully qualified container image names). Documentation
is also updated to describe how to set up for podman.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>