config: Update to latest tagged component versions
We prefer to default to a tagged component version rather than tracking
the tip of a branch, since this minimizes chances of hitting stability
issues (users that want to follow a branch can override it themselves).
However this means we need to periodically up-rev all components and
test that they continue to work. All tests have passed with these new
versions.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>