build, clean: Squelch noisy output from git and dtc
Git is chatty by default and so would previously output informtive
messages even when shrinkwrap's --verbose was not specified. This would
make the output less legible. So pass --quiet to git to shut if up when
the user has not specified --verbose.

Similarly, when injecting a CHOSEN node into a device tree, dtc would
re-emit warnings that are already output from the original compile.
Suppress these duplicates by passing -q to dtc when doing the recompile.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>