docker: Build FVP into aarch64 docker image
Now that Arm have released a public build of the FVP_Base_RevC-2xAEMvA
for aarch64 linux systems, add that to the aarch64 docker image, just
like we do with x86_64. This brings aarch64 to feature parity with
x86_64.

Note that while we are using FVP 11.20 for aarch64, we are still using
18.16 for x86_64. 11.20 is the first available release for aarch64, but
a performance regression was seen with it on x86_64 so we are holding
off upgrading for x86_64 until the regression is fixed.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>