config: Add Arm CCA demonstrator config
Brings together a software stack to demonstrate Arm CCA running on FVP
in a four-world configuration. Includes TF-A in root world, Hafnium and
some demo secure partitions in secure world, RMM in realm world, and
Linux in Normal world.

In order to launch realm VMs, the user must bring their own rootfs that
contains a realm-aware kvmtool and a payload suitable for execution in a
realm. This could be kvm-unit-tests or a copy of the kernel running as
host (since it also supports the RSI guest additions).

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>