test: Fix failing bootwrapper tests
Bootwrapper runtime tests were previously failing due to bootwrapper's
PSCI implementation not shutting down the FVP on request. So the FVP
would just hang at shutdown and the test would timeout. It turns out
that the FVP has a mechanism where it can search for a specific piece of
text from a uart and will shut itself down when it sees it. So let's use
this mechanism to look for the "reboot: System halted" message that
Linux outputs on shutdown. This fixes the test timeouts.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>